Replaces the dungeon layer w/ original underdark.
The old underdark returns. Classic Lich Dungeon.

Until properly mapped in, dungeondeep.dmm holds all of silkies dungeon map stuff to be expanded upon (Which was on the last Z-layer by accident as it was their first time mapping) as the vast labyrinthine maze that is the dreamdark beneath. It's getting worked on before re-release.

They can find the gateways, but they wont lead anywhere until the map has been made to load on roundstart.

- Bath-house portals were removed from the Shrine of Elysius with planned placement elsewhere in the forest.

- The Undead Village is getting made to reflect the 'hedge' like nature of the fey influence beneath the city.